\name{plotDim}
\alias{plotDim}
\title{ Determine the Plot Layout}
\description{
  Determine the "optimal" layout for the plots of MCMB sequences based on the number of 
parameters.}

\usage{
plotDim(p)
}

\arguments{
  \item{p}{the number of parameters}
}

\value{
  Dimensions for plot layout.
}

\keyword{hplot}
