
\name{rqmcmb.ci}
\alias{rqmcmb.ci}
\title{Summary of RQMCMB sequences}
\description{Provide mean, standard error and condifence intervals using RQMCMB 
sequences. The confidence intervals are based on standard normal quantiles. }

\usage{
  rqmcmb.ci(obj, alpha=0.1)
}
\arguments{
  \item{obj}{The output object from rqmcmb() }
  \item{alpha}{Significance level, default alhpa=0.10 }
}

\keyword{regression}

